+++ de0_nano/software/imu_test/MPU6050.h
@@ -400,7 +400,7 @@ THE SOFTWARE.
 class MPU6050 {
     public:
         MPU6050();
-        MPU6050(uint8_t address);
+        MPU6050(uint16_t address);
 
         void initialize();
         bool testConnection();
